klar

+++ functions.js
@@ -3,7 +3,7 @@
 
 //logs in user
 async function loginUser(username, password) {
-    let url = "http://127.0.0.1:3000/api/login";
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/login";
 
     //const response = await fetch(url);
     //const data = await response.json();
@@ -29,7 +29,7 @@ async function loginUser(username, password) {
 
 //registrate user
 async function registrerUser(firstname, lastname, email, username, password) {
-    let url = "http://127.0.0.1:3000/api/register";
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/register";
     
     let user = {
         firstname: firstname,
@@ -54,7 +54,7 @@ async function registrerUser(firstname, lastname, email, username, password) {
 
 //Gets data from database
 async function getData() {
-    let url = "http://127.0.0.1:3000/api/menu";
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu";
     let token = window.sessionStorage.getItem("token");
 
     console.log("token sent: " + token);
@@ -71,7 +71,7 @@ async function getData() {
 };
 //gets menu items by category
 async function getMenuItemsByCategory(category) {
-    let url = "http://127.0.0.1:3000/api/menu?category=" + category;
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu?category=" + category;
     let token = window.sessionStorage.getItem("token");
 
     console.log("token sent: " + token);
@@ -89,7 +89,7 @@ async function getMenuItemsByCategory(category) {
 
 //gets a specific menu item
 async function getMenuItem(id) {
-    let url = "http://127.0.0.1:3000/api/menu";
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu";
 
     const response = await fetch(url);
     const data = await response.json();
@@ -109,7 +109,7 @@ async function getMenuItem(id) {
 //gets a specific menu item category
 async function getMenuCategory(type) {
     
-    let url = "http://127.0.0.1:3000/api/menu";
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu";
 
     const response = await fetch(url);
     const data = await response.json();
@@ -128,7 +128,7 @@ async function getMenuCategory(type) {
 
 //creates new menu item
 async function createMenuItem(name, type, description, prize) {
-    let url = "http://127.0.0.1:3000/api/menu";
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu";
     let token = window.sessionStorage.getItem("token");
 
     let menuItem = {
@@ -153,7 +153,7 @@ async function createMenuItem(name, type, description, prize) {
 
 //changes a menu item
 async function changeMenuItem(id, name, type, description, prize) {
-    let url = "http://127.0.0.1:3000/api/menu/" + id;
+    let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu/" + id;
     let token = window.sessionStorage.getItem("token");
     
     let menuItem = {
@@ -183,23 +183,34 @@ async function changeMenuItem(id, name, type, description, prize) {
 
 //deletes a menu item
 async function deleteMenuItem(id) {
-    let url = "http://127.0.0.1:3000/api/menu/" + id;
-    let token = window.sessionStorage.getItem("token");
+    if (confirm("Confirm that you want to delete the selected item?")) {
+        let url = "https://whale-app-kmmcn.ondigitalocean.app/api/menu/" + id;
+        let token = window.sessionStorage.getItem("token");
+        
+        const response = await fetch(url, {
+            method: "DELETE",
+            headers: {
+                "content-type": "Application/json",
+                "Authorization": "Bearer " + token
+            },
+           
+        });
     
-    const response = await fetch(url, {
-        method: "DELETE",
-        headers: {
-            "content-type": "Application/json",
-            "Authorization": "Bearer " + token
-        },
-       
-    });
-
-    if (!response.ok) {
-        throw new Error("somethig went wrong");
+        if (!response.ok) {
+            throw new Error("somethig went wrong");
+        }
+    
+        window.location.reload();
+    } else {
+        return;
     }
-
-    window.location.reload();
 }
+
+
+function logout() {
+    window.sessionStorage.removeItem("token");
+    window.location.replace("login.html");
+}
+
